added category model and creteCategory functionality with role based access

+++ src/app/modules/user/user.service.ts
@@ -1,5 +1,8 @@
+import httpStatus from 'http-status';
+import ApiError from '../../../errors/ApiError';
 import prisma from '../../../shared/prisma';
 import { IUpdateUserResponse, IUserResponse } from './user.interface';
+import { checkUserExistency } from './user.utils';
 
 const getAllUser = async (): Promise<IUserResponse[]> => {
   const result = await prisma.user.findMany({
@@ -11,6 +14,8 @@ const getAllUser = async (): Promise<IUserResponse[]> => {
       contactNo: true,
       address: true,
       profileImg: true,
+      createdAt: true,
+      updatedAt: true,
     },
   });
   return result;
@@ -29,8 +34,13 @@ const getSingleUser = async (id: string): Promise<IUserResponse | null> => {
       contactNo: true,
       address: true,
       profileImg: true,
+      createdAt: true,
+      updatedAt: true,
     },
   });
+  if (!result) {
+    throw new ApiError(httpStatus.BAD_REQUEST, 'User not found');
+  }
   return result;
 };
 
@@ -38,6 +48,10 @@ const updateSingleUser = async (
   id: string,
   payload: IUpdateUserResponse
 ): Promise<IUpdateUserResponse | null> => {
+  const isExists = await checkUserExistency(id);
+  if (!isExists) {
+    throw new ApiError(httpStatus.BAD_REQUEST, 'User not found');
+  }
   const result = await prisma.user.update({
     where: {
       id,
@@ -50,12 +64,18 @@ const updateSingleUser = async (
       contactNo: true,
       address: true,
       profileImg: true,
+      createdAt: true,
+      updatedAt: true,
     },
   });
   return result;
 };
 
 const deleteSingleUser = async (id: string): Promise<IUserResponse | null> => {
+  const isExists = await checkUserExistency(id);
+  if (!isExists) {
+    throw new ApiError(httpStatus.BAD_REQUEST, 'User not found');
+  }
   const result = await prisma.user.delete({
     where: {
       id,
@@ -68,6 +88,8 @@ const deleteSingleUser = async (id: string): Promise<IUserResponse | null> => {
       contactNo: true,
       address: true,
       profileImg: true,
+      createdAt: true,
+      updatedAt: true,
     },
   });
   return result;
